reorganize assessment subsections into their own component directories
